requirements.txt
streamlit
chromadb
langchain
langchain-community
langchain-huggingface
sentence-transformers
pdfplumber
pytesseract
Pillow
llama-cpp-python
tiktoken
numpy
pandas
huggingface_hub
groq
python-dotenv
langchain-text-splitters
langchain-core
